Bug fix: Using NaN to keep track of the end of the sort is fragile.
"Clearing" the array elements using NaN is fragile; it works on doubles,
but not on integers (the NaN change to "0" during casting to integers).
Instead, declare the loop index "j" outside the loop, and detect if its
value is -1 at the end, meaning the loop ended without hitting the "break"
statement.